Fix duration bug
Previously, if the video duration was 60 seconds, it would bde returned from the youtube API as PT1M. Since the app was slicing [3:4] and coverting to integeter, this was breaking everything. Condition implemented to check and "M" at index 3 of the duration string before slicing. If the "M" exists, it defaults to 60.